Move public API tests to integration tests
